+++ src/Core/Operators/IAddOperator.cs
@@ -36,7 +36,7 @@ namespace Reko.Core.Operators
 		{
             if (!ValidArgs(c1, c2))
                 return Constant.Invalid;
-            return BuildConstant(c1.DataType, c2.DataType, (int) (c1.ToInt64() + c2.ToInt64()));
+            return BuildConstant(c1.DataType, c2.DataType, c1.ToInt64() + c2.ToInt64());
 		}
 
 		public override string ToString()
